+++ Source/Function/Merge.ts
@@ -2,9 +2,6 @@
  * @module Merge
  *
  */
-export default (await import("deepmerge-ts")).deepmergeCustom<Generic>({
+export default (await import("deepmerge-ts")).deepmergeCustom({
 	mergeArrays: false,
-}) satisfies Interface<Generic> as Interface<Generic>;
-
-import type Interface from "../Interface/Merge.js";
-import type { Generic } from "../Interface/Merge.js";
+});