+++ Source/Variable/CSS/csso.ts
@@ -5,7 +5,7 @@
 export default {
 	comments: false,
 	forceMediaMerge: true,
-	restructure: false,
-} satisfies Interface;
+	restructure: false
+} satisfies Interface as Interface;
 
 import type Interface from "../../Interface/CSS/csso.js";