+++ Source/Variable/CSS/lightningcss.ts
@@ -4,6 +4,7 @@
  */
 export default {
 	minify: true,
-} satisfies Interface;
+	sourceMap: false,
+} satisfies Interface as Interface;
 
 import type Interface from "../../Interface/CSS/lightningcss.js";