+++ Target/Function/Integration.d.ts
@@ -7,16 +7,27 @@ declare const _default: Interface;
 export default _default;
 import type Interface from "../Interface/Integration.js";
 import type Action from "@playform/pipe/Target/Interface/Action.js";
-export declare const Default: Omit<{} & {
+export declare const Default: {
+    Path: string;
+    Cache: {
+        Search: string;
+        Folder: string;
+    };
+    Logger: 2;
+    Action: {
+        Failed: ({ Input }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
+        Passed: ({ Before, Buffer }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<boolean>;
+        Accomplished: ({ Input, Before, After }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
+        Changed: (Plan: import("@playform/pipe/Target/Interface/Plan.js").default) => Promise<any>;
+        Read: ({ Input }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
+        Wrote: ({ Buffer }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<import("@playform/pipe/Target/Type/Buffer.js").Type>;
+        Fulfilled: ({ File }: import("@playform/pipe/Target/Interface/Plan.js").default) => Promise<string | false>;
+    };
+    File: string;
+    Exclude: false;
     CSS: {
-        csso: {
-            comments: false;
-            forceMediaMerge: true;
-            restructure: false;
-        };
-        lightningcss: {
-            minify: true;
-        };
+        csso: import("../Interface/CSS/csso.js").default;
+        lightningcss: import("../Interface/CSS/lightningcss.js").default;
     };
     HTML: {
         "html-minifier-terser": {
@@ -48,43 +59,7 @@ export declare const Default: Omit<{} & {
         };
     };
     Image: {
-        sharp: {
-            avif: {
-                chromaSubsampling: string;
-                effort: number;
-                lossless: true;
-            };
-            gif: {
-                effort: number;
-            };
-            jpeg: {
-                chromaSubsampling: string;
-                mozjpeg: true;
-                trellisQuantisation: true;
-                overshootDeringing: true;
-                optimiseScans: true;
-            };
-            png: {
-                compressionLevel: number;
-                palette: true;
-            };
-            tiff: {
-                compression: string;
-            };
-            webp: {
-                effort: number;
-                lossless: true;
-            };
-            heif: {
-                effort: number;
-                lossless: true;
-            };
-            sharp: {
-                failOn: "error";
-                sequentialRead: true;
-                unlimited: true;
-            };
-        };
+        sharp: import("../Interface/Image/sharp.js").default;
     };
     JavaScript: {
         terser: {
@@ -111,38 +86,19 @@ export declare const Default: Omit<{} & {
             plugins: "preset-default"[];
         };
     };
-    Map: {
-        CSS: string;
-        HTML: string;
-        Image: string;
-        JavaScript: string;
-        SVG: string;
-    };
-    Parser: {
-        CSS: ("csso" | "lightningcss")[];
-        HTML: "html-minifier-terser";
-        Image: "sharp";
-        JavaScript: "terser";
-        SVG: "svgo";
-    };
-    Path: string;
-    Cache: {
-        Search: string;
-        Folder: string;
-    };
-    Logger: 2;
-    Action: Omit<{} & {
-        Failed: ({ Input }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
-        Passed: ({ Before, Buffer }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<boolean>;
-        Accomplished: ({ Input, Before, After }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
-        Changed: (Plan: import("@playform/pipe/Target/Interface/Plan.js").default) => Promise<any>;
-        Read: ({ Input }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<string>;
-        Wrote: ({ Buffer }: import("@playform/pipe/Target/Interface/File.js").default) => Promise<import("@playform/pipe/Target/Type/Buffer.js").Type>;
-        Fulfilled: ({ File }: import("@playform/pipe/Target/Interface/Plan.js").default) => Promise<string | false>;
-    }, "__proto__">;
-    File: string;
-    Exclude: false;
-}, "__proto__">;
+    Map: import("../Interface/Map.js").default;
+    Parser: import("../Interface/Parser.js").default;
+};
 export declare const Search: string;
-export declare const Merge: import("../Interface/Merge.js").default<import("../Interface/Merge.js").Generic>;
+export declare const Merge: <Ts extends readonly unknown[]>(...objects: Ts) => import("deepmerge-ts").DeepMergeHKT<Ts, Readonly<{
+    DeepMergeRecordsURI: "DeepMergeRecordsDefaultURI";
+    DeepMergeArraysURI: "DeepMergeArraysDefaultURI";
+    DeepMergeSetsURI: "DeepMergeSetsDefaultURI";
+    DeepMergeMapsURI: "DeepMergeMapsDefaultURI";
+    DeepMergeOthersURI: "DeepMergeLeafURI";
+    DeepMergeFilterValuesURI: "DeepMergeFilterValuesDefaultURI";
+}>, Readonly<{
+    key: PropertyKey;
+    parents: readonly Readonly<Record<PropertyKey, unknown>>[];
+}>>;
 export declare let _Action: Action;