+++ Target/Variable/Parser.d.ts
@@ -2,11 +2,6 @@
  * @module Parser
  *
  */
-declare const _default: {
-    CSS: ("csso" | "lightningcss")[];
-    HTML: "html-minifier-terser";
-    Image: "sharp";
-    JavaScript: "terser";
-    SVG: "svgo";
-};
+declare const _default: Interface;
 export default _default;
+import type Interface from "../Interface/Parser.js";