+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## 2.2.23
+## 0.0.6
+
+-   Bug fix
+-   Implements lightningcss
+
+## 0.0.5
+
+-   Bug fix
+-   Implements lightningcss
+
+## 0.0.4
 
 -   Cleanup